pyproject.toml
[project]
name = "kentiq-banking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-speech>=1.49.1",
    "fastapi==0.115.0",
    "google-generativeai>=0.8.6",
    "gtts==2.5.3",
    "numpy==1.26.4",
    "openai==1.51.0",
    "openai-whisper>=20250625",
    "opencv-python==4.10.0.84",
    "pdfplumber==0.11.4",
    "pillow==10.4.0",
    "pyaudio==0.2.14",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "pyttsx3==2.90",
    "reportlab==4.2.2",
    "requests==2.32.3",
    "scipy==1.14.1",
    "sounddevice==0.5.0",
    "speechrecognition==3.10.4",
    "streamlit>=1.56.0",
    "uvicorn[standard]==0.30.6",
]
